ILEX-15 history item is moved to seperate file
